==> jupyter_notebooks/04_feature_engineering.ipynb <==
# **Feature Engineering Notebook**

Part of CRISP-DM **Data Preparation**

## Objectives

* Engineering features for Regression model

## Inputs

* outputs/datasets/cleaned/train_set_cleaned.csv
* outputs/datasets/cleaned/test_set_cleaned.csv
## Outputs

* Generate a list with variables to engineer



---

# Change working directory

We need to change the working directory from its current folder to its parent folder
* We access the current directory with os.getcwd()

import os
current_dir = os.getcwd()
current_dir

We want to make the parent of the current directory the new current directory
* os.path.dirname() gets the parent directory
* os.chir() defines the new current directory

os.chdir(os.path.dirname(current_dir))
print("You set a new current directory")

Confirm the new current directory

current_dir = os.getcwd()
current_dir

# Load Cleaned Data

Train Set

import pandas as pd
train_set_path = "outputs/datasets/cleaned/train_set_cleaned.csv"
TrainSet = pd.read_csv(train_set_path)
TrainSet.head(3)

Test Set

test_set_path = "outputs/datasets/cleaned/test_set_cleaned.csv"
TestSet = pd.read_csv(test_set_path)
TestSet.head(3)

---

# Data Exploration

Since we are interested in potentially transforming and optimizing features we start by generating a profile report.

from ydata_profiling import ProfileReport
pandas_report = ProfileReport(df=TrainSet, minimal=True)
pandas_report.to_notebook_iframe()

When checking the variables we note that we have several zeros, outliers and skewed distributions.
For our model to work we will also need to address the text variables.

# Feature Engineering

# # Code from walkthrough project 02, feature engineering notebook

import scipy.stats as stats
import matplotlib.pyplot as plt
import seaborn as sns
import pandas as pd
import warnings
from feature_engine import transformation as vt
from feature_engine.outliers import Winsorizer
from feature_engine.encoding import OrdinalEncoder
sns.set(style="whitegrid")
warnings.filterwarnings('ignore')


def FeatureEngineeringAnalysis(df, analysis_type=None):
    """
    - used for quick feature engineering on numerical and categorical variables
    to decide which transformation can better transform the distribution shape
    - Once transformed, use a reporting tool, like ydata-profiling, to evaluate distributions
    """
    check_missing_values(df)
    allowed_types = ['numerical', 'ordinal_encoder', 'outlier_winsorizer']
    check_user_entry_on_analysis_type(analysis_type, allowed_types)
    list_column_transformers = define_list_column_transformers(analysis_type)

    # Loop in each variable and engineer the data according to the analysis type
    df_feat_eng = pd.DataFrame([])
    for column in df.columns:
        # create additional columns (column_method) to apply the methods
        df_feat_eng = pd.concat([df_feat_eng, df[column]], axis=1)
        for method in list_column_transformers:
            df_feat_eng[f"{column}_{method}"] = df[column]

        # Apply transformers in respective column_transformers
        df_feat_eng, list_applied_transformers = apply_transformers(
            analysis_type, df_feat_eng, column)

        # For each variable, assess how the transformations perform
        transformer_evaluation(
            column, list_applied_transformers, analysis_type, df_feat_eng)

    return df_feat_eng


def check_user_entry_on_analysis_type(analysis_type, allowed_types):
    """ Check analysis type """
    if analysis_type is None:
        raise SystemExit(
            f"You should pass analysis_type parameter as one of the following options: {allowed_types}")
    if analysis_type not in allowed_types:
        raise SystemExit(
            f"analysis_type argument should be one of these options: {allowed_types}")


def check_missing_values(df):
    if df.isna().sum().sum() != 0:
        raise SystemExit(
            f"There is a missing value in your dataset. Please handle that before getting into feature engineering.")


def define_list_column_transformers(analysis_type):
    """ Set suffix columns according to analysis_type"""
    if analysis_type == 'numerical':
        list_column_transformers = [
            "log_e", "log_10", "reciprocal", "power", "box_cox", "yeo_johnson"]

    elif analysis_type == 'ordinal_encoder':
        list_column_transformers = ["ordinal_encoder"]

    elif analysis_type == 'outlier_winsorizer':
        list_column_transformers = ['iqr']

    return list_column_transformers


def apply_transformers(analysis_type, df_feat_eng, column):
    for col in df_feat_eng.select_dtypes(include='category').columns:
        df_feat_eng[col] = df_feat_eng[col].astype('object')

    if analysis_type == 'numerical':
        df_feat_eng, list_applied_transformers = FeatEngineering_Numerical(
            df_feat_eng, column)

    elif analysis_type == 'outlier_winsorizer':
        df_feat_eng, list_applied_transformers = FeatEngineering_OutlierWinsorizer(
            df_feat_eng, column)

    elif analysis_type == 'ordinal_encoder':
        df_feat_eng, list_applied_transformers = FeatEngineering_CategoricalEncoder(
            df_feat_eng, column)

    return df_feat_eng, list_applied_transformers


def transformer_evaluation(column, list_applied_transformers, analysis_type, df_feat_eng):
    # For each variable, assess how the transformations perform
    print(f"* Variable Analyzed: {column}")
    print(f"* Applied transformation: {list_applied_transformers} \n")
    for col in [column] + list_applied_transformers:

        if analysis_type != 'ordinal_encoder':
            DiagnosticPlots_Numerical(df_feat_eng, col)

        else:
            if col == column:
                DiagnosticPlots_Categories(df_feat_eng, col)
            else:
                DiagnosticPlots_Numerical(df_feat_eng, col)

        print("\n")


def DiagnosticPlots_Categories(df_feat_eng, col):
    plt.figure(figsize=(4, 3))
    sns.countplot(data=df_feat_eng, x=col, palette=[
                  '#432371'], order=df_feat_eng[col].value_counts().index)
    plt.xticks(rotation=90)
    plt.suptitle(f"{col}", fontsize=30, y=1.05)
    plt.show()
    print("\n")


def DiagnosticPlots_Numerical(df, variable):
    fig, axes = plt.subplots(1, 3, figsize=(12, 4))
    sns.histplot(data=df, x=variable, kde=True, element="step", ax=axes[0])
    stats.probplot(df[variable], dist="norm", plot=axes[1])
    sns.boxplot(x=df[variable], ax=axes[2])

    axes[0].set_title('Histogram')
    axes[1].set_title('QQ Plot')
    axes[2].set_title('Boxplot')
    fig.suptitle(f"{variable}", fontsize=30, y=1.05)
    plt.tight_layout()
    plt.show()


def FeatEngineering_CategoricalEncoder(df_feat_eng, column):
    list_methods_worked = []
    try:
        encoder = OrdinalEncoder(encoding_method='arbitrary', variables=[
                                 f"{column}_ordinal_encoder"])
        df_feat_eng = encoder.fit_transform(df_feat_eng)
        list_methods_worked.append(f"{column}_ordinal_encoder")

    except Exception:
        df_feat_eng.drop([f"{column}_ordinal_encoder"], axis=1, inplace=True)

    return df_feat_eng, list_methods_worked


def FeatEngineering_OutlierWinsorizer(df_feat_eng, column):
    list_methods_worked = []

    # Winsorizer iqr
    try:
        disc = Winsorizer(
            capping_method='iqr', tail='both', fold=1.5, variables=[f"{column}_iqr"])
        df_feat_eng = disc.fit_transform(df_feat_eng)
        list_methods_worked.append(f"{column}_iqr")
    except Exception:
        df_feat_eng.drop([f"{column}_iqr"], axis=1, inplace=True)

    return df_feat_eng, list_methods_worked


def FeatEngineering_Numerical(df_feat_eng, column):
    list_methods_worked = []

    # LogTransformer base e
    try:
        lt = vt.LogTransformer(variables=[f"{column}_log_e"])
        df_feat_eng = lt.fit_transform(df_feat_eng)
        list_methods_worked.append(f"{column}_log_e")
    except Exception:
        df_feat_eng.drop([f"{column}_log_e"], axis=1, inplace=True)

    # LogTransformer base 10
    try:
        lt = vt.LogTransformer(variables=[f"{column}_log_10"], base='10')
        df_feat_eng = lt.fit_transform(df_feat_eng)
        list_methods_worked.append(f"{column}_log_10")
    except Exception:
        df_feat_eng.drop([f"{column}_log_10"], axis=1, inplace=True)

    # ReciprocalTransformer
    try:
        rt = vt.ReciprocalTransformer(variables=[f"{column}_reciprocal"])
        df_feat_eng = rt.fit_transform(df_feat_eng)
        list_methods_worked.append(f"{column}_reciprocal")
    except Exception:
        df_feat_eng.drop([f"{column}_reciprocal"], axis=1, inplace=True)

    # PowerTransformer
    try:
        pt = vt.PowerTransformer(variables=[f"{column}_power"])
        df_feat_eng = pt.fit_transform(df_feat_eng)
        list_methods_worked.append(f"{column}_power")
    except Exception:
        df_feat_eng.drop([f"{column}_power"], axis=1, inplace=True)

    # BoxCoxTransformer
    try:
        bct = vt.BoxCoxTransformer(variables=[f"{column}_box_cox"])
        df_feat_eng = bct.fit_transform(df_feat_eng)
        list_methods_worked.append(f"{column}_box_cox")
    except Exception:
        df_feat_eng.drop([f"{column}_box_cox"], axis=1, inplace=True)

    # YeoJohnsonTransformer
    try:
        yjt = vt.YeoJohnsonTransformer(variables=[f"{column}_yeo_johnson"])
        df_feat_eng = yjt.fit_transform(df_feat_eng)
        list_methods_worked.append(f"{column}_yeo_johnson")
    except Exception:
        df_feat_eng.drop([f"{column}_yeo_johnson"], axis=1, inplace=True)

    return df_feat_eng, list_methods_worked

### Feature Engineering Summary

We will use the following transformers:
* Categorical Encoding
* Numerical Transformation
* Handeling of outliers using the Windsorizer
* Smart Correlation Selection

### Categorical Encoding
**Ordinal Encoding**: replaces categories with ordinal numbers

* Step 1: We select our categorical variables

variables_engineering= ['BsmtExposure', 'BsmtFinType1', 'GarageFinish', 'KitchenQual']

* Step 2: We create a separate DataFrame, with our variables

df_engineering = TrainSet[variables_engineering].copy()
df_engineering.head(3)

* Step 3: We create engineered variables by applying the transformation, assess engineered variables distribution and select the most suitable method for each variable.

df_engineering = FeatureEngineeringAnalysis(df=df_engineering, analysis_type='ordinal_encoder')

* For all variables, our transformation has effectively converted the categories to numbers.

* Step 4 - We apply the selected transformation to the Train and Test set

# the steps are: 
# 1 - create a transformer
# 2 - fit_transform into TrainSet
# 3 - transform into TestSet 
encoder = OrdinalEncoder(encoding_method='arbitrary', variables = variables_engineering)
TrainSet = encoder.fit_transform(TrainSet)
TestSet = encoder.transform(TestSet)

print("* Categorical encoding - ordinal transformation done!")

### Numerical Transformation

* Step 1: We select our numerical variables

variables_engineering = ['1stFlrSF',
                        '2ndFlrSF',
                        'BsmtFinSF1',
                        'BsmtUnfSF',
                        'GarageArea',
                        'GrLivArea',
                        'LotArea',
                        'LotFrontage',
                        'MasVnrArea',
                        'OpenPorchSF',
                        'TotalBsmtSF']

* Step 2: We create a separate DataFrame, with our variables

df_engineering = TestSet[variables_engineering].copy()
df_engineering.head(3)

* Step 3: We create engineered variables by applying the transformation, assess engineered variables distribution and select the most suitable method for each variable.

df_engineering = FeatureEngineeringAnalysis(df=df_engineering, analysis_type='numerical')

* Analyzing the plors we decide on the following:
    * We use log transformation on: '1stFlrSF', 'GrLivArea', 'LotArea', 'LotFrontage'
    * We use the power transformarion on: 'BsmtUnfSF', 'MasVnrArea', 'OpenPorchSF', 'TotalBsmtSF'
    * We use the yeo-johnson transformation on: 'GarageArea'
    * We use no transformation on: '2ndFlrSF', 'BsmtFinSF1', 'TotalBsmtSF'

* Step 4 - We apply the selected transformation to the Train and Test set

# the steps are: 
# 1 - create a transformer
# 2 - fit_transform into TrainSet
# 3 - transform into TestSet 
variables_engineering_log = ['1stFlrSF', 'GrLivArea', 'LotArea', 'LotFrontage']
variables_engineering_pow = ['BsmtUnfSF', 'MasVnrArea','OpenPorchSF','TotalBsmtSF']
variables_engineering_yeo = ['GarageArea']

log_trans = vt.LogTransformer(variables = variables_engineering_log, base='e')
pow_trans = vt.PowerTransformer(variables = variables_engineering_pow)
yeo_trans = vt.YeoJohnsonTransformer(variables = variables_engineering_yeo)

transformers = [log_trans, pow_trans, yeo_trans]
for t in transformers:
    TrainSet = t.fit_transform(TrainSet)
    TestSet = t.transform(TestSet)

print("* Numerical transformation done!")

### Winsoriser

We assess the outliers

df_engineering = TrainSet.copy()
df_engineering.head(3)

df_engineering = FeatureEngineeringAnalysis(df=df_engineering, analysis_type='outlier_winsorizer')

We conclude that the variables that benefit from handling the outliers are: 'GrLivArea', 'OpenPorchSF', 'TotalBsmtSF'.
We apply the transformation to these variables of the train and test sets.

# the steps are: 
# 1 - create a transformer
# 2 - fit_transform into TrainSet
# 3 - transform into TestSet 
variables_engineering= ['GrLivArea', 'OpenPorchSF', 'TotalBsmtSF']

winsoriser = Winsorizer(capping_method='iqr', tail='both', fold=1.5, variables = variables_engineering)
TrainSet = winsoriser.fit_transform(TrainSet)
TestSet = winsoriser.transform(TestSet)

print("* Outlier transformation done!")

### SmartCorrelatedSelection Variables

For this transformer there is no need to select variables, all the variables will be fed into the transformer. Since our target variable should not be included in this analysis we drop 'SalePrice'.

df_engineering = TrainSet.copy().drop(['SalePrice'],axis=1)
df_engineering.head(3)

from feature_engine.selection import SmartCorrelatedSelection
corr_sel = SmartCorrelatedSelection(variables=None, method="pearson", threshold=0.6, selection_method="variance")

corr_sel.fit_transform(df_engineering)
print("* The following groups of correlated variables have been identified:")
corr_sel.correlated_feature_sets_

print("* The Smart Correlation Selection has determined to drop the following features:")
corr_sel.features_to_drop_

# Feature Engineering Pipeline

from sklearn.pipeline import Pipeline


variables_engineering_ord= ['BsmtExposure','BsmtFinType1','GarageFinish','KitchenQual']
variables_engineering_log = ['1stFlrSF', 'GrLivArea', 'LotArea', 'LotFrontage']
variables_engineering_pow = ['BsmtUnfSF', 'MasVnrArea','OpenPorchSF','TotalBsmtSF']
variables_engineering_yeo = ['GarageArea']
variables_engineering_wins= ['GrLivArea', 'OpenPorchSF', 'TotalBsmtSF']

pipeline = Pipeline([
    ('ordinal_encoder', OrdinalEncoder(encoding_method='arbitrary', variables = variables_engineering_ord)),
    ('log_transformer', vt.LogTransformer(variables=variables_engineering_log,base='e')),
    ('power_transformer', vt.PowerTransformer(variables=variables_engineering_pow)),
    ('yeo_johnson_transformer', vt.YeoJohnsonTransformer(variables=variables_engineering_yeo)),
    ('windsorizer', Winsorizer(capping_method='iqr', tail='both', fold=1.5, variables=variables_engineering_wins)),
    ("SmartCorrelatedSelection", SmartCorrelatedSelection(variables=None,
        method="spearman", threshold=0.6, selection_method="variance")),
])

TrainSet = pipeline.fit_transform(TrainSet)
TestSet = pipeline.transform(TestSet)

# Conclusion

The table below shows the transformations that will be added to the ML Pipeline.

| Variable      | Proposed Transformation                                          |
|---------------|------------------------------------------------------------------|
| 1stFlrSF      | Numerical - log_e, SmartCorrelatedSelection                      |
| 2ndFlrSF      | SmartCorrelatedSelection                                         |
| BedroomAbvGr  | SmartCorrelatedSelection                                         |
| BsmtExposure  | OrdinalEncoder, SmartCorrelatedSelection                         |
| BsmtFinSF1    | SmartCorrelatedSelection                                         |
| BsmtFinType1  | OrdinalEncoder, SmartCorrelatedSelection                         |
| BsmtUnfSF     | Numerical - power, SmartCorrelatedSelection                |
| GarageArea    | Numerical - log_e, SmartCorrelatedSelection                |
| GarageFinish  | OrdinalEncoder, SmartCorrelatedSelection                         |
| GarageYrBlt   | SmartCorrelatedSelection                                         |
| GrLivArea     | Numerical - log_e, Winsorizer, SmartCorrelatedSelection         |
| KitchenQual   | OrdinalEncoder, SmartCorrelatedSelection                         |
| LotArea       | Numerical - log_e, SmartCorrelatedSelection                      |
| LotFrontage   | Numerical - log_e, SmartCorrelatedSelection                                         |
| MasVnrArea    | Numerical - power, SmartCorrelatedSelection                                         |
| OpenPorchSF   | Numerical - power, Winsoriser, SmartCorrelatedSelection                                         |
| OverallCond   | SmartCorrelatedSelection                                         |
| OverallQual   | SmartCorrelatedSelection                                         |
| SalePrice     | None - target variable                                           |
| TotalBsmtSF   | Winsoriser, SmartCorrelatedSelection                |
| YearBuilt     | SmartCorrelatedSelection                                         |
| YearRemodAdd  | SmartCorrelatedSelection                                         |
